Improve -n option. Start to use Template to generate html
The -n option is used to avoid get new data from the repository
and run the test, using only data stored in the colectors json file
to allow faster test of the process, when making changes to
html or js. This commit improve that functionality.

Also start to use Template to make easier to understand the html generation.